Port member tests to the new harness, etc.
Rename json-test/HelloWorld to "code" so all the tree names in the tests are the same and we can just hard-code "code" in the search URL search_results() uses.

This lets us remove the "*member*" and "BitField.h" files from json-test. Now we're getting somewhere.
