--- tests/json-test/HelloWorld/member_function.cpp
@@ -1,9 +1,0 @@
-class MemberFunction
-{
-public:
-    void member_function();
-};
-
-void MemberFunction::member_function()
-{
-}
