--- tests/json-test/HelloWorld/member_variable.cpp
@@ -1,5 +1,0 @@
-class MemberVariable
-{
-public:
-    int member_variable;
-};
