--- tests/json-test/HelloWorld/static_member.cpp
@@ -1,3 +1,0 @@
-#include "static_member.h"
-
-int StaticMember::static_member = 0;
