--- tests/json-test/HelloWorld/static_member.h
@@ -1,5 +1,0 @@
-class StaticMember
-{
-public:
-    static int static_member;
-};
