+++ tests/json-test/code/macro_if_defined.cpp
