+++ tests/json-test/code/macro_undef.cpp
