+++ tests/json-test/code/main.c
@@ -1,9 +1,7 @@
 #include <stdio.h>
-#include "BitField.h"
 
 // Hello World Example
 int main(int argc, char* argv[]){
   printf("Hello World");
-  BitField b(8);
   return 0;
 }
