+++ tests/json-test/code/prototype_parameter.cpp
