+++ tests/json-test/code/typedef.cpp
